Refactored and commented
